refactor: shared ui 컴포넌트 파일 분리
