+++ src/shared/ui/Dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './Dialog';
