+++ src/shared/ui/HighlightText/HighlightText.tsx
@@ -1,0 +1,22 @@
+interface HighlightTextProps {
+  text: string;
+  highlight: string;
+}
+
+export const HighlightText = ({ text, highlight }: HighlightTextProps) => {
+  if (!text) return null;
+  if (!highlight.trim()) {
+    return <span>{text}</span>;
+  }
+
+  const regex = new RegExp(`(${highlight})`, 'gi');
+  const parts = text.split(regex);
+
+  return (
+    <span>
+      {parts.map((part, i) =>
+        regex.test(part) ? <mark key={i}>{part}</mark> : <span key={i}>{part}</span>,
+      )}
+    </span>
+  );
+};
