+++ src/shared/ui/index.ts
@@ -1,0 +1,8 @@
+export * from './Card';
+export * from './Input';
+export * from './Dialog';
+export * from './Select';
+export * from './Table';
+export * from './Textarea';
+export * from './HighlightText';
+export * from './Button';
